+++ spec/rubocop/config_spec.rb
@@ -6,7 +6,7 @@ require "open3"
 
 describe "Rubocop default config" do
   def run_rubocop(path)
-    cops_path = File.expand_path("lib/anycable/rails/compatibility/rubocop.rb", PROJECT_ROOT)
+    cops_path = File.expand_path("lib/anycable/rails/rubocop.rb", PROJECT_ROOT)
     output, _status = Open3.capture2(
       "bundle exec rubocop --force-default-config -d -r #{cops_path} #{path}",
       chdir: File.join(__dir__, "fixtures")